Work on window-manager and update-manager
This probably won't even compile (or it will and be super buggy)
