+++ src/main/modules/update-manager.js
@@ -1,11 +1,13 @@
 import { autoUpdater } from "electron-updater";
+import { ipcMain } from "electron";
 import { ipcBroadcast } from "./ipc";
 
 // TODO: Lotsa docs
 
 export default class updateManager {
-  constructor(windowMan) {
-    this.windowManager = windowMan;
+  constructor(windowMan, configMan) {
+    this._windowManager = windowMan;
+    this._configManager = configMan;
 
     this.versionDetails = {
       "versions": process.versions,
@@ -32,12 +34,21 @@ export default class updateManager {
     "https://raw.githubusercontent.com/Hexellate/tracking/master/autoupdate-test/updates",
       "useMultipleRangeRequest": false,
     };
+
+    this._registerListeners();
   }
 
+  /**
+   *
+   * @param {import("electron").BrowserWindow} recipient The target of the version info
+   */
   ipcSendVer(recipient) {
     recipient.send("getVer", this.versionDetails);
   }
 
+  /**
+   * @return {import("builder-util-runtime").GenericServerOptions} The configured provider
+   */
   get provider() {
     const provider = this.providerBase;
     if (this.versionDetails.appVersion.prerelease === []) {
@@ -50,6 +61,14 @@ export default class updateManager {
     return provider;
   }
 
+  get updateDetails() {
+    return this._updateDetails;
+  }
+
+  get versionDetails() {
+    return this._versionDetails;
+  }
+
   initialize() {
     // eslint-disable-next-line global-require
     autoUpdater.logger = require("electron-log");
@@ -60,9 +79,34 @@ export default class updateManager {
   }
 
   _registerListeners() {
+    ipcMain.on("getVer", (event) => {
+      this.ipcSendVer(event.sender);
+    });
+
+    ipcMain.on("getUpd", (event) => {
+      autoUpdater.checkForUpdates().then((prom) => {
+        this.updateDetails.promise = prom;
+        this.updateDetails.cancellationToken = this.updateDetails.promise.cancellationToken;
+        this.ipcSendVer(event.sender);
+      });
+    });
+
+    ipcMain.on("dlUpd", () => {
+      autoUpdater.downloadUpdate().then((prom) => {
+        this.updateDetails.promise = prom;
+        this.updateDetails.hasdlpromisereturned = true;
+      });
+      this.updateDetails.status = "progressing";
+      ipcBroadcast({ "channel": "updaterChangeStatus", "windows": this._windowManager.windows });
+    });
+
+    ipcMain.on("installUpdate", () => {
+      autoUpdater.quitAndInstall();
+    });
+
     autoUpdater.on("checking-for-update", () => {
       this.updateDetails.status = "checking";
-      ipcBroadcast({ "channel": "updaterChangeStatus", "windows": this.windowManager.windows });
+      ipcBroadcast({ "channel": "updaterChangeStatus", "windows": this._windowManager.windows });
     });
 
     autoUpdater.on("update-available", (info) => {
